// frontend/src/types/generated/EmailTriageConditionValue.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type EmailTriageConditionValue = { "text": string } | { "bool": boolean } | { "string_list": Array<string> } | { "number": bigint } | { "money_cents": bigint } | { "date": string } | { "header": { name: string, value: string, } } | "empty";